requirements.txt
fastmcp
notion-client
python-dotenv
